+++ base/deprecated.jl
@@ -901,8 +901,8 @@ for f in (
         # base/special/erf.jl
         :erfcx, :erfi, :dawson,
         # base/special/bessel.jl
-        :airyprime, :airyai, :airyaiprime, :airybi, :airybiprime,
-        :airy, :airyx, :besselj0, :besselj1, :bessely0, :bessely1,
+        :airyai, :airyaiprime, :airybi, :airybiprime,
+        :besselj0, :besselj1, :bessely0, :bessely1,
         # base/math.jl
         :cbrt, :sinh, :cosh, :tanh, :atan, :asinh, :exp, :erf, :erfc, :exp2,
         :expm1, :exp10, :sin, :cos, :tan, :asin, :acos, :acosh, :atanh,
@@ -965,7 +965,7 @@ for f in (
         # base/special/gamma.jl
         :polygamma, :zeta, :beta, :lbeta,
         # base/special/bessel.jl
-        :airy, :airyx, :besseli, :besselix, :besselj, :besseljx,
+        :besseli, :besselix, :besselj, :besseljx,
         :besselk, :besselkx, :bessely, :besselyx, :besselh,
         :besselhx, :hankelh1, :hankelh2, :hankelh1x, :hankelh2x,
         # base/math.jl
@@ -1167,5 +1167,66 @@ for (dep, f, op) in [(:sumabs!, :sum!, :abs),
         ($f)($op, r, A; init=init)
     end
 end
+
+@deprecate airy(z::Number) airyai(z)
+@deprecate airyx(z::Number) airyaix(z)
+@deprecate airyprime(z::Number) airyaiprime(z)
+@deprecate airy{T<:Number}(x::AbstractArray{T}) airyai.(x)
+@deprecate airyx{T<:Number}(x::AbstractArray{T}) airyaix.(x)
+@deprecate airyprime{T<:Number}(x::AbstractArray{T}) airyprime.(x)
+
+function _airy(k::Integer, z::Complex128)
+    depwarn("`airy(k,x)` is deprecated, use `airyai(x)`, `airyaiprime(x)`, `airybi(x)` or `airybiprime(x)` instead.",:airy)
+    id = Int32(k==1 || k==3)
+    if k == 0 || k == 1
+        return _airy(z, id, Int32(1))
+    elseif k == 2 || k == 3
+        return _biry(z, id, Int32(1))
+    else
+        throw(ArgumentError("k must be between 0 and 3"))
+    end
+end
+function _airyx(k::Integer, z::Complex128)
+    depwarn("`airyx(k,x)` is deprecated, use `airyaix(x)`, `airyaiprimex(x)`, `airybix(x)` or `airybiprimex(x)` instead.",:airyx)
+    id = Int32(k==1 || k==3)
+    if k == 0 || k == 1
+        return _airy(z, id, Int32(2))
+    elseif k == 2 || k == 3
+        return _biry(z, id, Int32(2))
+    else
+        throw(ArgumentError("k must be between 0 and 3"))
+    end
+end
+
+for afn in (:airy,:airyx)
+    _afn = Symbol("_"*string(afn))
+    suf  = string(afn)[5:end]
+    @eval begin
+        function $afn(k::Integer, z::Complex128)
+            depwarn("`$afn(k,x)` is deprecated, use `airyai$suf(x)`, `airyaiprime$suf(x)`, `airybi$suf(x)` or `airybiprime$suf(x)` instead.",$(QuoteNode(afn)))
+            $_afn(k,z)
+        end
+
+        $afn(k::Integer, z::Complex) = $afn(k, float(z))
+        $afn{T<:AbstractFloat}(k::Integer, z::Complex{T}) = throw(MethodError($afn,(k,z)))
+        $afn(k::Integer, z::Complex64) = Complex64($afn(k, Complex128(z)))
+        $afn(k::Integer, x::Real) = $afn(k, float(x))
+        $afn(k::Integer, x::AbstractFloat) = real($afn(k, complex(x)))
+
+        function $afn{T<:Number}(k::Number, x::AbstractArray{T})
+            depwarn("`$afn(k::Number,x::AbstractArray)` is deprecated, use `airyai$suf.(x)`, `airyaiprime$suf.(x)`, `airybi$suf.(x)` or `airybiprime$suf.(x)` instead.",$(QuoteNode(afn)))
+            $_afn.(k,x)
+        end
+        function $afn{S<:Number}(k::AbstractArray{S}, x::Number)
+            depwarn("`$afn(k::AbstractArray,x::AbstractArray)` is deprecated, use `airyai$suf.(x)`, `airyaiprime$suf.(x)`, `airybi$suf.(x)` or `airybiprime$suf.(x)` instead.",$(QuoteNode(afn)))
+            $_afn.(k,x)
+        end
+        function $afn{S<:Number,T<:Number}(k::AbstractArray{S}, x::AbstractArray{T})
+            depwarn("`$afn(k::AbstractArray,x::AbstractArray)` is deprecated, use `airyai$suf.(x)`, `airyaiprime$suf.(x)`, `airybi$suf.(x)` or `airybiprime$suf.(x)` instead.",$(QuoteNode(afn)))
+            $_afn.(k,x)
+        end
+    end
+end
+
 
 # End deprecations scheduled for 0.6